Get rid of obviated throughput measure
